requirements.txt
Django==2.0.5
Pillow==5.1.0
celery==4.1.1
colorific==0.3
django_celery_beat==1.1.1
django-extra-views==0.11.0
ipdb==0.11
ipython==6.4.0
python-dateutil==2.7.3
requests==2.18.4